Discord bot bugfixes

+++ lib/teiserver/bridge/chat_commands.ex
@@ -1,6 +1,6 @@
 defmodule Teiserver.Bridge.ChatCommands do
   @moduledoc false
-  alias Teiserver.{Account, User, Communication}
+  alias Teiserver.{Account, User, Communication, Config}
   alias Teiserver.Data.Types, as: T
   alias Teiserver.Bridge.UnitNames
   alias Nostrum.Api
@@ -35,57 +35,51 @@ defmodule Teiserver.Bridge.ChatCommands do
   end
 
   def handle_command({_user, _discord_id, message_id}, "gdt", _remaining, channel_id) do
-    gdt_forum =
-      Application.get_env(:central, DiscordBridge)[:bridges]
-      |> Enum.filter(fn {_, name} -> name == "gdt-discussion" end)
+    gdt_discussion_channel_id = Config.get_site_config_cache("teiserver.Discord channel #gdt-discussion")
 
-    case gdt_forum do
-      [{forum_id, _}] ->
-        # Post message to channel
-        Api.create_message(
-          channel_id,
-          "Thank you for your suggestion, the game design team will be discussing it. Once they have finished discussing it they will vote on it and post an update to this thread."
-        )
+    if gdt_discussion_channel_id do
+      # Post message to channel
+      Api.create_message(
+        channel_id,
+        "Thank you for your suggestion, the game design team will be discussing it. Once they have finished discussing it they will vote on it and post an update to this thread."
+      )
 
-        # Delete the message that was posted
-        Api.delete_message(channel_id, message_id)
+      # Delete the message that was posted
+      Api.delete_message(channel_id, message_id)
 
-        # channel_id = 1071140326644920353
-        {:ok, channel} = Api.get_channel(channel_id)
+      # channel_id = 1071140326644920353
+      {:ok, channel} = Api.get_channel(channel_id)
 
-        # Create new thread in gdt-discussion
-        {:ok, thread} =
-          Api.start_thread(forum_id, %{
-            name: "Discussion for #{channel.name}",
-            message: %{
-              content: "Thread to discuss #{channel.name} - <##{channel_id}>"
-            },
-            type: 11
-          })
-
-        {:ok, message} =
-          Api.create_message(thread.id, %{
+      # Create new thread in gdt-discussion
+      {:ok, thread} =
+        Api.start_thread(gdt_discussion_channel_id, %{
+          name: "Discussion for #{channel.name}",
+          message: %{
             content: "Thread to discuss #{channel.name} - <##{channel_id}>"
-          })
+          },
+          type: 11
+        })
 
-        # Pin message
-        Api.add_pinned_channel_message(thread.id, message.id)
+      {:ok, message} =
+        Api.create_message(thread.id, %{
+          content: "Thread to discuss #{channel.name} - <##{channel_id}>"
+        })
 
-        # Add GDTs to thread
-        Account.list_users(
-          search: [
-            gdt_member: "GDT"
-          ],
-          select: [:data]
-        )
-        |> Enum.map(fn %{data: data} -> data["discord_id"] end)
-        |> Enum.reject(&(&1 == nil))
-        |> Enum.each(fn user_discord_id ->
-          Nostrum.Api.add_thread_member(thread.id, user_discord_id)
-        end)
+      # Pin message
+      Api.add_pinned_channel_message(thread.id, message.id)
 
-      _ ->
-        :ok
+      # Add GDTs to thread
+      Account.list_users(
+        search: [
+          gdt_member: "GDT"
+        ],
+        select: [:data]
+      )
+      |> Enum.map(fn %{data: data} -> data["discord_id"] end)
+      |> Enum.reject(&(&1 == nil))
+      |> Enum.each(fn user_discord_id ->
+        Nostrum.Api.add_thread_member(thread.id, user_discord_id)
+      end)
     end
 
     :ignore
